requirements.txt
aiogram==3.10
Pillow>=10.0.0
python-dotenv>=1.0.0
aiojobs>=0.2.0
requests>=2.31.0
aiohttp>=3.9.0

